merge newest code

+++ person_task_manager/server/src/modules/person_task_manager/services/project.service.ts
@@ -1,44 +1,91 @@
 import { IResponse } from "../../../common/response";
-import { msg200 } from "../../../common/response_helpers";
+import { msg200, msg400 } from "../../../common/response_helpers";
 import { ProjectEntity } from "../entities/project.entity";
+import { projectValidation } from "../validations/project.validation";
+
+const projectValidationImpl = projectValidation;
 
 class ProjectService {
     constructor() {}
 
+    async createProject(project: any): Promise<IResponse> {
+        const createProject = await ProjectEntity.create(project);
+        
+        return msg200({
+            message: (createProject as any)
+        });
+    }
+
+    async updateProject(projectId: string, project: any): Promise<IResponse> {
+        try {
+            if (await projectValidationImpl.checkExistedProjectByProjectId(projectId)) {
+                const updateProject = await ProjectEntity.updateOne({_id: projectId}, project);
+                
+                return msg200({
+                    message: JSON.stringify(updateProject)
+                });             
+            } else {
+                return msg400("Project not found");
+            }
+        } catch (err: any) {
+            return msg400(err.message.toString())
+        }
+    }
+
+
+    async deleteProject(projectId: string): Promise<IResponse> {
+        try {
+            if (await projectValidationImpl.checkExistedProjectByProjectId(projectId)) {
+                const deleteProject = await ProjectEntity.deleteOne({_id: projectId});
+                
+                return msg200({
+                    message: JSON.stringify(deleteProject)
+                });
+            } else {
+                return msg400("Project not found");
+            }
+        } catch (err: any) {
+            return msg400(err.message.toString())
+        }
+    }
+
     async getProject(projectId: string): Promise<IResponse> {
         const project = await ProjectEntity.findOne({ _id: projectId });   
+        
         return msg200({
             project
         });
     }
 
     async getAllProjects(): Promise<IResponse> {
-        const projects = await ProjectEntity.find();
+        const projects = await ProjectEntity.find({ ownerId: 1});
+        
         return msg200({
             projects
         });
     }
 
-    async createProject(project: any): Promise<IResponse> {
-        const createProject = await ProjectEntity.create(project);
+    async getGroupTasksInProject(projectId: string): Promise<IResponse> {
+        const groupTasks = await ProjectEntity.findOne({_id: projectId}).populate('groupTasks');
+        
         return msg200({
-            message: (createProject as any).message
+            message: (groupTasks as any)
         });
     }
 
-    async updateProject(projectId: string, project: any): Promise<IResponse> {
-        const updateProject = await ProjectEntity.updateOne({_id: projectId}, project);
+    async updateManyProjects(data: any, project: any): Promise<IResponse> {
+        const updateManyProjects = await ProjectEntity.updateMany({data}, project);
+        
         return msg200({
-            message: (updateProject as any).message
+            message: (updateManyProjects as any)
         });
     }
 
-    async deleteProject(projectId: string): Promise<IResponse> {
-        const deleteProject = await ProjectEntity.deleteOne({_id: projectId});
-        return msg200({
-            message: (deleteProject as any).message
-        });
-    }
+    // disable project
+
+    // enable project
+
+    // archive project
 }
 
 export const projectService = new ProjectService();
